day22: solved part 2(still cleaning needed)

+++ src/main/java/be/durvenproeven/aoc2021/day22/Cube.java
@@ -33,10 +33,10 @@ public class Cube {
 				new Coordinates(xRange.getMax(), yRange.getMax(), zRange.getMax()));
 	}
 
-	public int getSize(){
+	public long getSize(){
 		return Arrays.stream(Axes.values())
 				.map(this::getIntegerRange)
-				.mapToInt(IntegerRange::getSize)
+				.mapToLong(IntegerRange::getSize)
 				.reduce(1, (a, b) -> a * b);
 	}
